split into an additional component
