+++ src/components/queries/queries-list.tsx
@@ -1,32 +1,21 @@
 import { FC, useCallback, useState } from 'react'
-import {
-  Alert,
-  AlertProps,
-  Box,
-  Button,
-  Snackbar,
-  TextField
-} from '@mui/material'
+import { Alert, AlertProps, Box, Button, Snackbar } from '@mui/material'
 import { DataGrid } from '@mui/x-data-grid'
 import {
-  useAddFlipQueryMutation,
   useDeletePoeFlipQueryMutation,
   useEditFlipQueryMutation,
   useGetPoeFlipQueryQuery
-} from '../../../lib/api-config'
+} from '../../../lib/apiConfig'
 import Link from '../../Link'
 import { createQueriesColumns } from './grid-colums-queries'
-import { FlipQueryType, NewFlipQueryType } from '../../types/flip-queries'
+import { FlipQueryTypes } from '../../types/flipQueryTypes'
+import { AddQueryFlip } from './add-query-flip'
 
-export const QueriesList: FC<{ queries: FlipQueryType[] }> = ({ queries }) => {
+export const QueriesList: FC<{ queries: FlipQueryTypes[] }> = ({ queries }) => {
   const { data: rows = queries } = useGetPoeFlipQueryQuery()
   const [removeQuery] = useDeletePoeFlipQueryMutation()
-  const [addQuery] = useAddFlipQueryMutation()
   const [editQuery] = useEditFlipQueryMutation()
-  const [newRow, setNewRow] = useState<NewFlipQueryType>({
-    cardQuery: '',
-    itemQuery: ''
-  })
+
   const [snackbar, setSnackbar] = useState<Pick<
     AlertProps,
     'children' | 'severity'
@@ -38,20 +27,8 @@ export const QueriesList: FC<{ queries: FlipQueryType[] }> = ({ queries }) => {
     setSnackbar({ children: err.message, severity: 'error' })
   }, [])
 
-  const addRowFromServerHandler = useCallback(
-    async () =>
-      addQuery(newRow)
-        .then(() => {
-          setSnackbar({ children: 'Successfully saved', severity: 'success' })
-        })
-        .catch((err) => {
-          setSnackbar({ children: `${err.data.error}`, severity: 'error' })
-        }),
-    [addQuery, newRow]
-  )
-
   const processRowUpdate = useCallback(
-    async (newGridRow: FlipQueryType, oldGridRow: FlipQueryType) =>
+    async (newGridRow: FlipQueryTypes, oldGridRow: FlipQueryTypes) =>
       editQuery(newGridRow)
         .unwrap()
         .then(() => {
@@ -71,7 +48,7 @@ export const QueriesList: FC<{ queries: FlipQueryType[] }> = ({ queries }) => {
   )
 
   const deleteRowHandlerFromServer = useCallback(
-    async (deletedRow: FlipQueryType) =>
+    async (deletedRow: FlipQueryTypes) =>
       removeQuery(deletedRow)
         .unwrap()
         .then(() => {
@@ -96,29 +73,7 @@ export const QueriesList: FC<{ queries: FlipQueryType[] }> = ({ queries }) => {
           go to main
         </Button>
       </Box>
-      <Box display="flex">
-        <TextField
-          sx={{ width: 499, marginLeft: 62 }}
-          id="cardQuery"
-          label="Card query"
-          variant="filled"
-          value={newRow.cardQuery}
-          onChange={(ev) =>
-            setNewRow({ ...newRow, cardQuery: ev.target.value })
-          }
-        />
-        <TextField
-          sx={{ width: 499, marginLeft: 1 }}
-          id="itemQuery"
-          label="Item query"
-          variant="filled"
-          value={newRow.itemQuery}
-          onChange={(ev) =>
-            setNewRow({ ...newRow, itemQuery: ev.target.value })
-          }
-        />
-        <Button onClick={addRowFromServerHandler}>Add new query</Button>
-      </Box>
+      <AddQueryFlip setSnackbar={setSnackbar} />
       <div>
         <DataGrid
           getRowId={(row) => JSON.parse(row.cardQuery).query?.type}
@@ -141,7 +96,7 @@ export const QueriesList: FC<{ queries: FlipQueryType[] }> = ({ queries }) => {
             horizontal: 'center'
           }}
           onClose={handleCloseSnackbar}
-          autoHideDuration={6000}
+          autoHideDuration={5000}
         >
           <Alert {...snackbar} onClose={handleCloseSnackbar} />
         </Snackbar>
